move common write mzml functions; minor refactoring
